add: merge settings file
